// app/server/applicationsDB.server/applications/dolphin/types.ts
export type DolphinButtonId =
  | "Buttons/A"
  | "Buttons/B"
  | "Buttons/X"
  | "Buttons/Y"
  | "Buttons/Z"
  | "Buttons/Start"
  | "Main Stick/Up"
  | "Main Stick/Down"
  | "Main Stick/Left"
  | "Main Stick/Right"
  | "Main Stick/Modifier"
  | "Main Stick/Calibration"
  | "C-Stick/Up"
  | "C-Stick/Down"
  | "C-Stick/Left"
  | "C-Stick/Right"
  | "C-Stick/Modifier"
  | "C-Stick/Calibration"
  | "Triggers/L"
  | "Triggers/R"
  | "Triggers/L-Analog"
  | "Triggers/R-Analog"
  | "D-Pad/Up"
  | "D-Pad/Down"
  | "D-Pad/Left"
  | "D-Pad/Right";